Finished editing basic storyboards
Basic edits are done. Both storyboards (iPhone and iPad) are working,
though there is one bug in the iPad version that causes the
overviewView to vanish unexpectedly.

Next up will be to see if it's possible to dynamically load the
different versions of the GLKViewController subclass into a single
storyboard. Otherwise, I'll have to load up a storybook for each one of
the GLKit examples, and that will be a hassle when so much of the data
is the same.